Remove conditions.py and move code into utils.py
