+++ psrqpy/__init__.py
@@ -21,4 +21,3 @@ __citation__ = """@article{psrqpy,
 from .search import QueryATNF
 from .pulsar import Pulsar, Pulsars
 from .utils import *
-from .conditions import *
